lioncrest-server/pyproject.toml
[project]
name = "lioncrest-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "colorlog>=6.9.0",
]
